Add copyright notices. Refs #1.
Add copyright notices and years, including a pointer and notice for the license
of the icons used by this application.

Notices are added to the LICENSE file, the cabal file, and all files in src/.

The package license is set to OtherLicense to indicate the fact that, even
though the license is permissive and open source, the icons are not BSD3.